Rename particle and animation related component
