+++ cocos/core/animation/animation-manager.ts
@@ -14,7 +14,7 @@ import { AnimationState } from './animation-state';
 import { CrossFade } from './cross-fade';
 import { legacyCC } from '../global-exports';
 import { IJointTransform, deleteTransform, getTransform, getWorldMatrix } from '../renderer/models/skinning-model';
-import { Socket } from './skeletal-animation-component';
+import { Socket } from './skeletal-animation';
 
 interface ISocketData {
     target: Node;
