+++ cocos/core/animation/index.ts
@@ -22,7 +22,7 @@ export {
 export * from './animation-component';
 export * from './skeletal-animation-data-hub';
 export * from './skeletal-animation-state';
-export * from './skeletal-animation-component';
+export * from './skeletal-animation';
 export * from './transform-utils';
 export { animation };
